Recycler View was added to the activity_main.xml and appropriate code was added for its functionality
The images of the products are using the android icon as the images at hand were too big and causing the app to misbehave when rendered.

A floatingAction Button was added to the same activity.

There is some minor UX problem in the main activity concerning the scrolling but can be overlooked for now.